requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
click==8.3.1
commonmark==0.9.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
mdpdf==0.0.18
openai==2.30.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.2.2
pypandoc==1.17
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==1.0.0
tomli==2.4.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
